// web/admin/gql/schema.tsx
export type PickArrayType<T> = T extends (infer U)[] ? U : NonNullable<T>
export const crs = { type: 'name', properties: { name: 'urn:ogc:def:crs:EPSG::4326' } }
export type Maybe<T> = T | null | undefined
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }

// Generated on 2021-10-04T01:13:08+05:30


/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any
  /** A universally unique identifier as defined by [RFC 4122](https://tools.ietf.org/html/rfc4122). */
  UUID: string
  /** A point in time as described by the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone. */
  Datetime: Date
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { [key: string]: unknown }
  /** A floating point number that requires more precision than IEEE 754 binary 64 */
  BigFloat: any
  /** The `GeoJSON` scalar type represents GeoJSON values as specified by[RFC 7946](https://tools.ietf.org/html/rfc7946). */
  GeoJSON: GeometryObject
  Email: string
  /** Relaxed phone number validation */
  PhoneUs: any
  /** The day, does not include a time. */
  Date: any
  /** The exact time of day, does not include the date. May or may not have a timezone offset. */
  Time: string
  /** A JSON Web Token defined by [RFC 7519](https://tools.ietf.org/html/rfc7519) which securely represents claims between two parties. */
  JwtToken: string
}
// ...
export enum SlotStatusT {
  Enabled = 'ENABLED',
  Disabled = 'DISABLED',
  Unlisted = 'UNLISTED',
  Deleted = 'DELETED',
}

/** Verification status for slots */
export enum SlotVerificationStatus {
  Pending = 'PENDING',
  Rejected = 'REJECTED',
  Verified = 'VERIFIED',
}
// ...
export enum AccessRestrictions {
  None = 'NONE',
  Barrier = 'BARRIER',
  Keycard = 'KEYCARD',
  Keycode = 'KEYCODE',
}

/** values for business_status are: active, inactive */
export enum SlotBusinessStatus {
  Active = 'ACTIVE',
  Inactive = 'INACTIVE',
}

/** Allowed values for slot_category are: private, public */
export enum SlotCategory {
  Private = 'PRIVATE',
  Public = 'PUBLIC',
}
// ...
export type SlotOwnerIdFkeyInput = {
  /** The primary key(s) for `user` for the far side of the relationship. */
  connectById?: Maybe<UserUserPkeyConnect>
  /** The primary key(s) for `user` for the far side of the relationship. */
  connectByNodeId?: Maybe<UserNodeIdConnect>
  /** The primary key(s) for `user` for the far side of the relationship. */
  deleteById?: Maybe<UserUserPkeyDelete>
  /** The primary key(s) for `user` for the far side of the relationship. */
  deleteByNodeId?: Maybe<UserNodeIdDelete>
  /** The primary key(s) and patch data for `user` for the far side of the relationship. */
  updateById?: Maybe<UserOnSlotForSlotOwnerIdFkeyUsingUserPkeyUpdate>
  /** The primary key(s) and patch data for `user` for the far side of the relationship. */
  updateByNodeId?: Maybe<SlotOnSlotForSlotOwnerIdFkeyNodeIdUpdate>
}
// ...
export type SlotVehicleSizeIdFkeyInput = {
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  connectById?: Maybe<VehicleSizeVehicleSizePkeyConnect>
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  connectByNodeId?: Maybe<VehicleSizeNodeIdConnect>
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  deleteById?: Maybe<VehicleSizeVehicleSizePkeyDelete>
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  deleteByNodeId?: Maybe<VehicleSizeNodeIdDelete>
  /** The primary key(s) and patch data for `vehicleSize` for the far side of the relationship. */
  updateById?: Maybe<VehicleSizeOnSlotForSlotVehicleSizeIdFkeyUsingVehicleSizePkeyUpdate>
  /** The primary key(s) and patch data for `vehicleSize` for the far side of the relationship. */
  updateByNodeId?: Maybe<SlotOnSlotForSlotVehicleSizeIdFkeyNodeIdUpdate>
  /** A `VehicleSizeInput` object that will be created and connected to this object. */
  create?: Maybe<SlotVehicleSizeIdFkeyVehicleSizeCreateInput>
}
// ...
export type SlotParkingSpaceIdFkeyInput = {
  /** The primary key(s) for `parkingSpace` for the far side of the relationship. */
  connectById?: Maybe<ParkingSpaceParkingSpacePkeyConnect>
  /** The primary key(s) for `parkingSpace` for the far side of the relationship. */
  connectBySlug?: Maybe<ParkingSpaceParkingSpaceSlugUkeyConnect>
  /** The primary key(s) for `parkingSpace` for the far side of the relationship. */
  connectByNodeId?: Maybe<ParkingSpaceNodeIdConnect>
  /** The primary key(s) for `parkingSpace` for the far side of the relationship. */
  deleteById?: Maybe<ParkingSpaceParkingSpacePkeyDelete>
  /** The primary key(s) for `parkingSpace` for the far side of the relationship. */
  deleteBySlug?: Maybe<ParkingSpaceParkingSpaceSlugUkeyDelete>
  /** The primary key(s) for `parkingSpace` for the far side of the relationship. */
  deleteByNodeId?: Maybe<ParkingSpaceNodeIdDelete>
  /** The primary key(s) and patch data for `parkingSpace` for the far side of the relationship. */
  updateById?: Maybe<ParkingSpaceOnSlotForSlotParkingSpaceIdFkeyUsingParkingSpacePkeyUpdate>
  /** The primary key(s) and patch data for `parkingSpace` for the far side of the relationship. */
  updateBySlug?: Maybe<ParkingSpaceOnSlotForSlotParkingSpaceIdFkeyUsingParkingSpaceSlugUkeyUpdate>
  /** The primary key(s) and patch data for `parkingSpace` for the far side of the relationship. */
  updateByNodeId?: Maybe<SlotOnSlotForSlotParkingSpaceIdFkeyNodeIdUpdate>
  /** A `ParkingSpaceInput` object that will be created and connected to this object. */
  create?: Maybe<SlotParkingSpaceIdFkeyParkingSpaceCreateInput>
}
// ...
export type SlotAmenitySlotIdFkeyInverseInput = {
  /** Flag indicating whether all other `slotAmenity` records that match this relationship should be removed. */
  deleteOthers?: Maybe<Scalars['Boolean']>
  /** The primary key(s) for `slotAmenity` for the far side of the relationship. */
  connectById?: Maybe<Array<SlotAmenitySlotAmenityPkeyConnect>>
  /** The primary key(s) for `slotAmenity` for the far side of the relationship. */
  connectByNodeId?: Maybe<Array<SlotAmenityNodeIdConnect>>
  /** The primary key(s) for `slotAmenity` for the far side of the relationship. */
  deleteById?: Maybe<Array<SlotAmenitySlotAmenityPkeyDelete>>
  /** The primary key(s) for `slotAmenity` for the far side of the relationship. */
  deleteByNodeId?: Maybe<Array<SlotAmenityNodeIdDelete>>
  /** The primary key(s) and patch data for `slotAmenity` for the far side of the relationship. */
  updateById?: Maybe<
    Array<SlotAmenityOnSlotAmenityForSlotAmenitySlotIdFkeyUsingSlotAmenityPkeyUpdate>
  >
  /** The primary key(s) and patch data for `slotAmenity` for the far side of the relationship. */
  updateByNodeId?: Maybe<Array<SlotOnSlotAmenityForSlotAmenitySlotIdFkeyNodeIdUpdate>>
  /** A `SlotAmenityInput` object that will be created and connected to this object. */
  create?: Maybe<Array<SlotAmenitySlotIdFkeySlotAmenityCreateInput>>
}
// ...
export type SlotAvailabilitySlotIdFkeyInverseInput = {
  /** Flag indicating whether all other `slotAvailability` records that match this relationship should be removed. */
  deleteOthers?: Maybe<Scalars['Boolean']>
  /** The primary key(s) for `slotAvailability` for the far side of the relationship. */
  connectById?: Maybe<Array<SlotAvailabilitySlotAvailabilityPkeyConnect>>
  /** The primary key(s) for `slotAvailability` for the far side of the relationship. */
  connectByNodeId?: Maybe<Array<SlotAvailabilityNodeIdConnect>>
  /** The primary key(s) for `slotAvailability` for the far side of the relationship. */
  deleteById?: Maybe<Array<SlotAvailabilitySlotAvailabilityPkeyDelete>>
  /** The primary key(s) for `slotAvailability` for the far side of the relationship. */
  deleteByNodeId?: Maybe<Array<SlotAvailabilityNodeIdDelete>>
  /** The primary key(s) and patch data for `slotAvailability` for the far side of the relationship. */
  updateById?: Maybe<
    Array<
      SlotAvailabilityOnSlotAvailabilityForSlotAvailabilitySlotIdFkeyUsingSlotAvailabilityPkeyUpdate
    >
  >
  /** The primary key(s) and patch data for `slotAvailability` for the far side of the relationship. */
  updateByNodeId?: Maybe<Array<SlotOnSlotAvailabilityForSlotAvailabilitySlotIdFkeyNodeIdUpdate>>
  /** A `SlotAvailabilityInput` object that will be created and connected to this object. */
  create?: Maybe<Array<SlotAvailabilitySlotIdFkeySlotAvailabilityCreateInput>>
}
// ...
export type CreateSlotInput = {
  /** An arbitrary string value with no semantic meaning. Will be included in the payload verbatim. May be used to track mutations by the client. */
  clientMutationId?: Maybe<Scalars['String']>
  /** The `Slot` to be created by this mutation. */
  slot: SlotInput
}

/** An input for mutations affecting `Slot` */
export type SlotInput = {
  id?: Maybe<Scalars['UUID']>
  name: Scalars['String']
  ownerId?: Maybe<Scalars['UUID']>
  vehicleSizeId?: Maybe<Scalars['UUID']>
  address?: Maybe<Scalars['JSON']>
  timezone?: Maybe<Scalars['String']>
  pricePerHour: Scalars['BigFloat']
  status?: Maybe<SlotStatusT>
  photoUrl?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
  notes?: Maybe<Scalars['String']>
  slug?: Maybe<Scalars['String']>
  createdAt?: Maybe<Scalars['Datetime']>
  updatedAt?: Maybe<Scalars['Datetime']>
  deleted?: Maybe<Scalars['Boolean']>
  deletedAt?: Maybe<Scalars['Datetime']>
  verificationStatus?: Maybe<SlotVerificationStatus>
  parkingSpaceId?: Maybe<Scalars['UUID']>
  location: Scalars['GeoJSON']
  shape?: Maybe<Scalars['GeoJSON']>
  accessRestrictions?: Maybe<AccessRestrictions>
  businessStatus?: Maybe<SlotBusinessStatus>
  businessStatusReason?: Maybe<Scalars['String']>
  category?: Maybe<SlotCategory>
  contributorId?: Maybe<Scalars['UUID']>
  level?: Maybe<Scalars['Int']>
  mapSource?: Maybe<Scalars['UUID']>
  slotDimensions?: Maybe<Scalars['JSON']>
  tempUnavailable?: Maybe<Scalars['Boolean']>
  tempUnavailableFrom?: Maybe<Scalars['Datetime']>
  tempUnavailableTo?: Maybe<Scalars['Datetime']>
  waypoints?: Maybe<Scalars['JSON']>
  userToOwnerId?: Maybe<SlotOwnerIdFkeyInput>
  vehicleSizeToVehicleSizeId?: Maybe<SlotVehicleSizeIdFkeyInput>
  parkingSpaceToParkingSpaceId?: Maybe<SlotParkingSpaceIdFkeyInput>
  slotAmenitiesUsingId?: Maybe<SlotAmenitySlotIdFkeyInverseInput>
  slotAvailabilitiesUsingId?: Maybe<SlotAvailabilitySlotIdFkeyInverseInput>
  slotBookingsUsingId?: Maybe<FakeApiSlotBookingsForeignKey0InverseInput>
}
// ...
export type CreateSlotMutation = {
  createSlot?: Maybe<{
    slot?: Maybe<{
      id: string
      name: string
      timezone: string
      slug?: Maybe<string>
      accessRestrictions?: Maybe<AccessRestrictions>
      parkingSpaceId?: Maybe<string>
      level?: Maybe<number>
      notes?: Maybe<string>
      address?: Maybe<{ [key: string]: unknown }>
      tempUnavailable?: Maybe<boolean>
      tempUnavailableFrom?: Maybe<Date>
      tempUnavailableTo?: Maybe<Date>
      location: { longitude: number; latitude: number }
      slotAmenitiesList: Array<{ amenity?: Maybe<{ id: string; name: string }> }>
      slotAvailabilitiesList: Array<{
        id: string
        dayOfWeek: number
        startHour: string
        endHour: string
      }>
      shape?: Maybe<{ geojson?: Maybe<GeometryObject> }>
    }>
  }>
}
// ...
export const CreateSlotDocument = gql`
  mutation CreateSlot($payload: CreateSlotInput!) {
    createSlot(input: $payload) {
      slot {
        id
        name
        timezone
        slug
        accessRestrictions
        parkingSpaceId
        level
        location {
          longitude: x
          latitude: y
        }
        notes
        address
        slotAmenitiesList {
          amenity {
            id
            name
          }
        }
        slotAvailabilitiesList {
          id
          dayOfWeek
          startHour
          endHour
        }
        shape {
          geojson
        }
        tempUnavailable
        tempUnavailableFrom
        tempUnavailableTo
      }
    }
  }
`